chore: refactor parse boolish
refactor parse boolish, moved into env_vars and additional convenience
method to read env var and treat as bool
